requirements.txt
Biopython
rich